+++ NMEA_Parser/NMEA_Parser.cpp
@@ -69,7 +69,7 @@ nmea_dgps_t NMEA_Parser::dgps() {
 	return _dgps;
 }
 
-double NMEA_Parser::hdop() {
+float NMEA_Parser::hdop() {
 	return _hdop;
 }
 
@@ -87,6 +87,36 @@ nmea_fix_quality_t NMEA_Parser::fix_quality() {
 
 bool NMEA_Parser::fix() {
 	return _fix;
+}
+
+nmea_err_t NMEA_Parser::store(nmea_storage_t* storage) {
+	storage->__timestamp          = _timestamp;
+	storage->__coordinates        = _coordinates;
+	storage->__altitude_sealevel  = _alt_sea;
+	storage->__altitude_wgs84     = _alt_wgs;
+	storage->__velocity           = _velocity;
+	storage->__dgps               = _dgps;
+	storage->__hdop               = _hdop;
+	storage->__magnetic_variation = _magvar;
+	storage->__satellites_visible = _satellites_visible;
+	storage->__fix_quality        = _fix_quality;
+	storage->__fix                = _fix;
+	return nmea_success;
+}
+
+nmea_err_t NMEA_Parser::load(nmea_storage_t* storage) {
+	_timestamp          = storage->__timestamp;
+	_coordinates        = storage->__coordinates;
+	_alt_sea            = storage->__altitude_sealevel;
+	_alt_wgs            = storage->__altitude_wgs84;
+	_velocity           = storage->__velocity;
+	_dgps               = storage->__dgps;
+	_hdop               = storage->__hdop;
+	_magvar             = storage->__magnetic_variation;
+	_satellites_visible = storage->__satellites_visible;
+	_fix_quality        = storage->__fix_quality;
+	_fix                = storage->__fix;
+	return nmea_success;
 }
 
 #ifdef ARDUINO
@@ -114,9 +144,9 @@ void NMEA_Parser::print_info() {
 	Serial.print("Fix Quality: ");
 	Serial.println((uint8_t)_fix_quality);
 	Serial.print("Location: ");
-	Serial.print((double)_coordinates.latitude / 100.0);
+	Serial.print((float)_coordinates.latitude / 100.0);
 	Serial.print(", ");
-	Serial.println((double)_coordinates.longitude / 100.0);
+	Serial.println((float)_coordinates.longitude / 100.0);
 	Serial.print("Altitude (sea level): ");
 	Serial.println(_alt_sea);
 	Serial.print("Altitude (WGS84):     ");
@@ -261,7 +291,7 @@ nmea_err_t NMEA_Parser::parse_gga(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_err_baddata;
 	}
-	err = parse_double(nmea, &_hdop);
+	err = parse_float(nmea, &_hdop);
 	if (err != nmea_success) {
 		return err;
 	}
@@ -271,7 +301,7 @@ nmea_err_t NMEA_Parser::parse_gga(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_err_baddata;
 	}
-	err = parse_double(nmea, &_alt_sea);
+	err = parse_float(nmea, &_alt_sea);
 	if (err != nmea_success) {
 		return err;
 	}
@@ -285,7 +315,7 @@ nmea_err_t NMEA_Parser::parse_gga(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_err_baddata;
 	}
-	err = parse_double(nmea, &_alt_wgs);
+	err = parse_float(nmea, &_alt_wgs);
 	if (err != nmea_success) {
 		return err;
 	}
@@ -433,7 +463,7 @@ nmea_err_t NMEA_Parser::parse_rmc(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_err_baddata;
 	}
-	err = parse_double(nmea, &_velocity.speed);
+	err = parse_float(nmea, &_velocity.speed);
 	if (err != nmea_success) {
 		return err;
 	}
@@ -443,7 +473,7 @@ nmea_err_t NMEA_Parser::parse_rmc(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_err_baddata;
 	}
-	err = parse_double(nmea, &_velocity.heading);
+	err = parse_float(nmea, &_velocity.heading);
 	if (err != nmea_success) {
 		return err;
 	}
@@ -467,7 +497,7 @@ nmea_err_t NMEA_Parser::parse_rmc(char* nmea, uint8_t len) {
 	if (nmea == NULL) {
 		return nmea_success;
 	}
-	err = parse_double(nmea, &_magvar);
+	err = parse_float(nmea, &_magvar);
 	if (err == nmea_success) {
 		nmea = strchr(&nmea[1], ',');
 		if (nmea == NULL) {
@@ -616,7 +646,7 @@ nmea_err_t NMEA_Parser::parse_int(char* nmea, uint16_t* store) {
 	return nmea_success;
 }
 
-nmea_err_t NMEA_Parser::parse_double(char* nmea, double* store) {
+nmea_err_t NMEA_Parser::parse_float(char* nmea, float* store) {
 	*store = 0.0;
 	register uint8_t fracs = 0;
 	register bool in_fracs = false;
@@ -636,7 +666,7 @@ nmea_err_t NMEA_Parser::parse_double(char* nmea, double* store) {
 				continue;
 		}
 		*store *= 10.0;
-		*store += (double)(nmea[idx] - '0');
+		*store += (float)(nmea[idx] - '0');
 		if (in_fracs) {
 			++fracs;
 		}